requirements.txt
# Environment
python-dotenv>=1.0.0

# LangChain
langchain-core>=0.3.0
langchain-community>=0.3.0
langchain-ollama>=0.1.0

# Vector Store
chromadb>=0.5.0

# Document Processing
pypdf>=4.0.0

# UI
streamlit>=1.38.0

# Embeddings
sentence-transformers>=2.2.0